Begin refactoring commands
THIS VERSION WILL NOT COMPILE.
This is the beginning of an effort to refactor commands, as the Listener previously contained much more than the listener as a result of lazy design.
