+++ src/main/java/net/theinfinitymc/infinitybot/commands/Command.java
@@ -1,0 +1,10 @@
+package net.theinfinitymc.infinitybot.commands;
+
+import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
+
+public interface Command {
+
+	void execute(MessageReceivedEvent event, String[] args);
+
+	String getDescription();
+}
