+++ src/main/java/net/theinfinitymc/infinitybot/commands/Help.java
@@ -1,0 +1,17 @@
+package net.theinfinitymc.infinitybot.commands;
+
+import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
+
+public class Help implements Command {
+
+	@Override
+	public void execute(MessageReceivedEvent event, String[] args) {
+
+	}
+
+	@Override
+	public String getDescription() {
+		return null;
+	}
+
+}
