+++ src/main/java/net/theinfinitymc/infinitybot/commands/Noot.java
@@ -1,0 +1,17 @@
+package net.theinfinitymc.infinitybot.commands;
+
+import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
+
+public class Noot implements Command {
+
+	@Override
+	public void execute(MessageReceivedEvent event, String[] args) {
+		attemptAddToQueue("noot.mp3", e.getGuild(), e.getTextChannel(), e.getAuthor());
+	}
+
+	@Override
+	public String getDescription() {
+		return null;
+	}
+
+}
